// vitis-vision-front-end/src/app/core/api/models/block-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BlockResponse {
  id?: number;
  name?: string;
  partitioningType?: string;
  rowOrientation?: string;
  rowSpacing?: number;
  trellisSystemType?: string;
  vineSpacing?: number;
  createdAt?: string;
  createdBy?: string;
  lastUpdatedAt?: string;
  lastUpdatedBy?: string;
}